feat(eth): Enable ethernet in simulation. Update firmware and bootloader.
Also remove unused tester simulation files.
The only change in the tester verilator testbench in relation to iob-soc
was the 'tfp' file generation instead of 'vcd'. It would be better to
add a generic option in the iob-soc testbench to change this, instead of
rewriting the entire testbench here. Having a new testbench in this
repository makes it harder to maintain.